Move new MlsGroup to correct directories (#639)
